#567 - dynamic interface to Data Service endpoints
