+++ lib/endpoint-proxy.js
@@ -351,6 +351,253 @@ function paramMapEntry(paramDeclaration) {
   return [paramdef.name(), paramdef];
 }
 
+// START SHARED WITH Modules/MarkLogic/rest-api/lib/openapi-transform.sjs IN THE SERVER
+function expandDataDeclaration(id, dataDeclaration) {
+  const datatype = dataDeclaration.datatype;
+  if (datatype === void 0 || datatype === null) {
+    throw new Error(`missing datatype for ${id}`);
+  }
+
+  switch (datatype) {
+    case 'boolean':
+    case 'date':
+    case 'dateTime':
+    case 'dayTimeDuration':
+    case 'decimal':
+    case 'double':
+    case 'float':
+    case 'int':
+    case 'long':
+    case 'string':
+    case 'time':
+    case 'unsignedInt':
+    case 'unsignedLong':
+      dataDeclaration.dataKind = 'atomic';
+      dataDeclaration.mimeType = 'text/plain';
+      break;
+    case 'binaryDocument':
+      dataDeclaration.dataKind = 'node';
+      dataDeclaration.mimeType = 'application/x-unknown-content-type';
+      break;
+    case 'array':
+    case 'jsonDocument':
+    case 'object':
+      dataDeclaration.dataKind = 'node';
+      dataDeclaration.mimeType = 'application/json';
+      break;
+    case 'textDocument':
+      dataDeclaration.dataKind = 'node';
+      dataDeclaration.mimeType = 'text/plain';
+      break;
+    case 'xmlDocument':
+      dataDeclaration.dataKind = 'node';
+      dataDeclaration.mimeType = 'application/xml';
+      break;
+    default:
+      throw new Error(`invalid datatype configuration ${datatype} for ${id}`);
+  }
+
+  const multiple = dataDeclaration.multiple;
+  if (multiple === void 0 || multiple === null) {
+    dataDeclaration.multiple = false;
+  } else if (typeof multiple !== 'boolean' && !(multiple instanceof Boolean)) {
+    throw new Error(`invalid multiple configuration ${multiple} for ${id}`);
+  } else if (multiple === true && datatype === 'session') {
+    throw new Error(`session cannot be multiple for ${id}`);
+  }
+
+  const nullable = dataDeclaration.nullable;
+  if (nullable === void 0 || nullable === null) {
+    dataDeclaration.nullable = false;
+  } else if (typeof nullable !== 'boolean' && !(nullable instanceof Boolean)) {
+    throw new Error(`invalid nullable configuration ${nullable} for ${id}`);
+  }
+}
+function expandParamDeclaration(paramDeclaration) {
+  const paramName = paramDeclaration.name;
+  if (paramName === void 0 || paramName === null) {
+    throw new Error(`missing parameter name`);
+  }
+
+  expandDataDeclaration('${paramName} parameter', paramDeclaration);
+}
+function expandReturnDeclaration(returnDeclaration) {
+  expandDataDeclaration('return value', returnDeclaration);
+
+  let jstype = returnDeclaration.$jsType;
+  if (jstype === void 0) {
+    jstype = null;
+  }
+
+  const datatype = returnDeclaration.datatype;
+  switch (datatype) {
+    case 'boolean':
+      if (jstype === null) {
+        returnDeclaration.$jsType = 'boolean';
+      } else if (jstype !== 'boolean' && jstype !== 'string') {
+        throw new Error(
+            `optional $jsType ${jstype} can only be "boolean" or "string" for datatype ${datatype} return value`
+        );
+      }
+      break;
+    case 'date':
+    case 'dayTimeDuration':
+    case 'decimal':
+    case 'double':
+    case 'long':
+    case 'string':
+    case 'time':
+    case 'unsignedLong':
+      if (jstype === null) {
+        returnDeclaration.$jsType = 'string';
+      } else if (jstype !== 'string') {
+        throw new Error(
+            `optional $jsType ${jstype} can only be "string" for datatype ${datatype} return value`
+        );
+      }
+      break;
+    case 'dateTime':
+      if (jstype === null) {
+        returnDeclaration.$jsType = 'string';
+      } else if (jstype !== 'Date' && jstype !== 'string') {
+        throw new Error(
+            `optional $jsType ${jstype} can only be "Date" or "string" for datatype ${datatype} return value`
+        );
+      }
+      break;
+    case 'float':
+    case 'int':
+    case 'unsignedInt':
+      if (jstype === null) {
+        returnDeclaration.$jsType = 'number';
+      } else if (jstype !== 'number' && jstype !== 'string') {
+        throw new Error(
+            `optional $jsType ${jstype} can only be "number" or "string" for datatype ${datatype} return value`
+        );
+      }
+      break;
+    case 'array':
+    case 'binaryDocument':
+    case 'jsonDocument':
+    case 'object':
+    case 'textDocument':
+    case 'xmlDocument':
+      if (jstype !== null) {
+        throw new Error(`cannot specify a $jsType for datatype ${datatype} for return value`);
+      }
+      break;
+    default:
+      throw new Error(`unknown datatype ${datatype} for return value`);
+  }
+}
+function expandFunctionDeclaration(functionDeclaration) {
+  const functionName = functionDeclaration.functionName;
+  if (functionName === void 0 || functionName === null) {
+    throw new Error(`missing function name`);
+  }
+
+  let paramsKind   = 'empty';
+  let sessionParam = null;
+
+  const paramDeclarations = functionDeclaration.params;
+  if (paramDeclarations === void 0 || paramDeclarations === null) {
+    functionDeclaration.params = [];
+    functionDeclaration.maxArgs = 0;
+  } else if (paramDeclarations.length > 0) {
+    const filteredDeclarations = paramDeclarations.filter(paramDeclaration => {
+      if (paramDeclaration === void 0 || paramDeclaration === null) {
+        return false;
+      } else if (paramDeclaration.datatype === 'session') {
+        if (sessionParam !== null) {
+          throw new Error(
+              `${functionName} function with multiple session parameters: ${sessionParam.name} and ${paramDeclaration.name}`
+          );
+        }
+        sessionParam = paramDeclaration;
+        return false;
+      }
+
+      expandParamDeclaration(paramDeclaration);
+
+      switch (paramsKind) {
+        case 'empty':
+          const dataKind = paramDeclaration.dataKind;
+          switch (dataKind) {
+            case 'atomic':
+              paramsKind = 'multiAtomic';
+              break;
+            case 'node':
+              paramsKind = 'multiNode';
+              break;
+            default:
+              throw new Error(
+                  `invalid kind of param data ${dataKind} for ${paramDeclaration.name} parameter`
+              );
+          }
+          break;
+        case 'multiAtomic':
+          if (paramDeclaration.dataKind === 'node') {
+            paramsKind = 'multiNode';
+          }
+          break;
+        case 'multiNode':
+          break;
+        default:
+          throw new Error(
+              `invalid kind of parameter list ${paramsKind} for ${paramDeclaration.name} parameter`
+          );
+      }
+
+      return true;
+    });
+
+    if (filteredDeclarations.length < paramDeclarations.length) {
+      functionDeclaration.params  = filteredDeclarations;
+      functionDeclaration.maxArgs = (sessionParam !== null) ?
+          (filteredDeclarations.length + 1) : filteredDeclarations.length;
+    } else {
+      functionDeclaration.maxArgs = paramDeclarations.length;
+    }
+  }
+
+  functionDeclaration.paramsKind = paramsKind;
+
+  if (sessionParam !== null) {
+    sessionParam.dataKind = 'session';
+    sessionParam.mimeType = null;
+  }
+  functionDeclaration.sessionParam = sessionParam;
+
+  const returnDeclaration = functionDeclaration.return;
+  if (returnDeclaration === void 0) {
+    functionDeclaration.return = null;
+    functionDeclaration.returnKind = 'empty';
+  } else if (returnDeclaration === null) {
+    functionDeclaration.returnKind = 'empty';
+  } else {
+    expandReturnDeclaration(returnDeclaration);
+    functionDeclaration.returnKind = returnDeclaration.multiple ? 'multipart' : 'single';
+  }
+
+  const outputMode = functionDeclaration.$jsOutputMode;
+  if (outputMode === void 0 || outputMode === null) {
+    functionDeclaration.$jsOutputMode = 'promise';
+  } else if (outputMode !== 'promise' && outputMode !== 'stream') {
+    throw new Error(`${functionName} must have $jsOutputMode of "promise" or "stream" instead of ${outputMode}`);
+  }
+}
+function expandEndpointDeclaration(endpointDeclaration) {
+  const moduleExtension = endpointDeclaration.moduleExtension;
+  if (typeof moduleExtension !== "string" || moduleExtension.length === 0) {
+    throw new Error('invalid module extension configuration: ' + moduleExtension);
+  } else if (!moduleExtension.startsWith('.')) {
+    endpointDeclaration.moduleExtension = '.' + moduleExtension;
+  }
+
+  expandFunctionDeclaration(endpointDeclaration.declaration);
+}
+// END SHARED WITH Modules/MarkLogic/rest-api/lib/openapi-transform.sjs IN THE SERVER
+
 class Datadef {
   constructor(dataDeclaration) {
     this._dataKind = dataDeclaration.dataKind;
@@ -521,11 +768,92 @@ class EndpointProxy {
     }
   }
 }
+class ServiceCaller {
+  constructor(client, serviceDeclaration, endpointDeclarations) {
+    if (client === void 0 || client === null) {
+      throw new Error('no database client provided');
+    }
+    if (serviceDeclaration === void 0 || serviceDeclaration === null) {
+      throw new Error('no service declaration provided');
+    } else if (!(typeof serviceDeclaration.endpointExtension === 'string' || (serviceDeclaration.endpointExtension instanceof String))) {
+      throw new Error('serviceDeclaration.endpointExtension must be a String');
+    }
+
+    if(endpointDeclarations === null || (Array.isArray(endpointDeclarations) && endpointDeclarations.length > 0)) {
+      throw new Error('endpointDeclarations cannot be null or empty.');
+    }
+
+    const db = client.getConnectionParams().database;
+    if (db !== void 0 && db !== null) {
+      throw new Error('connection for client cannot specify database');
+    }
+
+    this.$mlProxy = endpointDeclarations.reduce((proxy,  endpointDeclarations) =>
+    {
+      expandFunctionDeclaration(endpointDeclarations);
+      return proxy.withFunction(endpointDeclarations, serviceDeclaration.endpointExtension);
+    }, client.createProxy(client, serviceDeclaration)
+  );
+  }
+
+  call(functionName, args) {
+    return this.$mlProxy.execute(functionName, args, 0);
+  }
+}
+
+class EndpointCaller {
+  constructor(client, endpointDeclaration) {
+    if (client === void 0 || client === null) {
+      throw new Error('no database client provided');
+    }
+    if (endpointDeclaration === void 0 || endpointDeclaration === null) {
+      throw new Error('endpointDeclaration cannot be null.');
+    }
+    const endpoint = endpointDeclaration.endpoint;
+
+    if (endpoint === void 0 || endpoint === null) {
+      throw new Error('endpointDeclaration endpoint cannot be null.');
+    }
+    const regex = /^(.*[/])([^/]+)(.[^/.]+)$/;
+    const arrayValues = endpoint.match(regex);
+
+    const endpointDirectory = arrayValues[1];
+    const endpointFunctionName = arrayValues[2];
+    const endpointExtension = arrayValues[3];
+
+    if(endpointDirectory == null || endpointExtension == null) {
+      throw new Error('endpoint must contain endpointDirectory and extension');
+    }
+    this.functionName = endpointFunctionName;
+    endpointDeclaration.functionName = this.functionName;
+
+    const serviceDeclaration = mlutil.copyProperties(endpointDeclaration);
+    serviceDeclaration.functionName = endpointFunctionName;
+
+    this.$mlProxy = client.createProxy(client, serviceDeclaration).withFunction(serviceDeclaration, endpointExtension);
+
+  }
+  call(args) {
+    return this.$mlProxy.execute(this.functionName, args, 0);
+  }
+}
 
 function makeEndpointProxy(dbClient, serviceDeclaration) {
   return new EndpointProxy(dbClient, serviceDeclaration);
 }
 
+function initService(dbClient, serviceDeclaration, endpointDeclarations) {
+  return new ServiceCaller(dbClient, serviceDeclaration, endpointDeclarations);
+}
+
+function initEndpoint(dbClient, endpointDeclaration) {
+  return new EndpointCaller(dbClient, endpointDeclaration);
+}
+
+
 module.exports = {
-  init: makeEndpointProxy
+  init: makeEndpointProxy,
+  expandEndpointDeclaration: expandEndpointDeclaration,
+  initService: initService,
+  initEndpoint: initEndpoint
 };
