+++ lib/marklogic.js
@@ -716,6 +716,20 @@ MarkLogicClient.prototype.createProxy = function createProxy(serviceDeclaration)
   return proxy.init(this, serviceDeclaration);
 };
 
+MarkLogicClient.prototype.serviceCaller = function serviceCaller(serviceDeclaration, endpointDeclarations) {
+  if(serviceDeclaration == null || endpointDeclarations == null) {
+    throw new Error('Parameters - serviceDeclaration and endpointDeclarations are required');
+  }
+  return proxy.initService(this, serviceDeclaration, endpointDeclarations);
+};
+
+MarkLogicClient.prototype.endpointCaller =function endpointCaller(endpointDeclaration) {
+  if(endpointDeclaration == null) {
+    throw new Error('endpointDeclaration cannot be null');
+  }
+  return proxy.initEndpoint(this, endpointDeclaration);
+};
+
 function initClient(client, inputParams) {
   var connectionParams = {};
   var isSSL = (inputParams.ssl == null) ? false : inputParams.ssl;
@@ -843,6 +857,8 @@ function prefLookup(hostname, options, callback) {
   };
   dns.lookup(hostname, {all:true, verbatim:false}, prefCallback);
 }
+
+
 
 module.exports = {
     createDatabaseClient: MarkLogicClient,
